requirements.txt
llvmlite==0.32.1
numba==0.49.1
numpy==1.18.4
Pillow==7.1.2
scipy==1.4.1
